[Feat] : 스티커 이미지 적용 및 static파일로 만듦
