+++ src/statics/stickers.json
@@ -1,0 +1,42 @@
+[
+  {
+    "id": 1,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_1.png"
+  },
+  {
+    "id": 2,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_2.png"
+  },
+  {
+    "id": 3,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_3.png"
+  },
+  {
+    "id": 4,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_4.png"
+  },
+  {
+    "id": 5,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_5.png"
+  },
+  {
+    "id": 6,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_6.png"
+  },
+  {
+    "id": 7,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_7.png"
+  },
+  {
+    "id": 8,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_8.png"
+  },
+  {
+    "id": 9,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_9.png"
+  },
+  {
+    "id": 10,
+    "src": "https://tooit-icon.s3.ap-northeast-2.amazonaws.com/sticker_10.png"
+  }
+]
